update vtune
vtune collection continues until the test app exits, so this script cannot kill the app to complete the collection. for a simple test, added an exe which computes the number of primes between 1 and 100,000,000 using OpenMP. each iteration of this takes ~1 sec on my machine, so I have it looping for ~30 seconds, then it finishes and vtune completes. still unable to generate a report due to some error